[CK DSL] Move gfx950 deep-conv-fusion case study into examples/gfx950/deep_conv_fusion
Collect the deep-fusion test scripts (deep_fused_conv_pool_verify,
compare_pool_tile_configs, profile_best_config) and the 11 analysis notes
into a single examples/gfx950/deep_conv_fusion package next to the other
gfx950 showcases. Fix the cross-imports and the python -m / file-path
references in the notes to the new module path.

No kernel behavior change: conv/transform edits are unaffected; conv parity
(implicit_gemm + direct_conv) and gfx950 GEMM verify still pass, and the
moved deep-fusion scripts verify bit-identical.
