+++ projects/composablekernel/python/ck_dsl/examples/gfx950/deep_conv_fusion/2026-06-02-1723-gfx950-roofline-bottlenecks.md
