+++ projects/composablekernel/python/ck_dsl/examples/gfx950/deep_conv_fusion/2026-06-03-0439-gfx950-conv1-lds-vectorization-results.md
@@ -141,12 +141,12 @@ Success signal for the next change is unchanged: `SQ_INSTS_LDS_STORE` falls and
 
 ```text
 # verify + bench
-<venv>/python -m ck_dsl.examples.gfx950.deep_fused_conv_pool_verify \
+<venv>/python -m ck_dsl.examples.gfx950.deep_conv_fusion.deep_fused_conv_pool_verify \
   --verify --bench --h 2160 --w 3840 --c 8 --k0 32 --k1 24 \
   --warmup 100 --iters 200
 
 # counters (same pmc groups as the 0406 baseline note)
 rocprofv3 -i pmc.txt -d <outdir> -o vec -f csv -- \
-  <venv>/python -m ck_dsl.examples.gfx950.deep_fused_conv_pool_verify \
+  <venv>/python -m ck_dsl.examples.gfx950.deep_conv_fusion.deep_fused_conv_pool_verify \
   --verify --h 2160 --w 3840 --c 8 --k0 32 --k1 24
 ```
